fix: refactor navigation
