+++ app/views/components/BountyForm.js
