+++ app/views/components/GroupListItem.js
