+++ app/views/components/SetBounty.js
