+++ app/views/components/ViewBounty.js
