+++ app/views/screens/Home.js
@@ -2,10 +2,10 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import { Container, Content, Text, Button, Form, Item, Icon, Input, List, ListItem, Body, Label, Title, Header } from 'native-base'
-import GroupListItem from '../../groupTab/components/GroupListItem'
-import { selectGroup } from '../../../redux/reducers/groups'
-import { setGroupUsers } from '../../../redux/reducers/users'
-import { setGroupTasks } from '../../../redux/reducers/users'
+import GroupListItem from '../components/GroupListItem'
+import { selectGroup } from '../../redux/reducers/groups'
+import { setGroupUsers } from '../../redux/reducers/users'
+import { setGroupTasks } from '../../redux/reducers/users'
 
 
 
@@ -47,6 +47,9 @@ class HomeComponent extends React.Component {
          <Button transparent onPress={this.toggleForm}>
            <Icon name="add-circle" />
            <Text>New Group</Text>
+         </Button>
+         <Button onPress={()=>{navigate('Profile')}}>
+           <Text>Temp Button to Profile</Text>
          </Button>
          {formState &&
          <Form>
