+++ app/views/styles/welcomeStyles.js
